S-61046 Created middleware/es6/deveoCommitMalformedError.js
Also, corrected DeveoTranslator propery issue in ES6. Should be "author",
not "committer". Somehow the ES6 had the old "committer" property.

The DeveoCommitMalformedError that was checked in also referred to GitLab
internally, so that is fixed now too.
