+++ src/app/api/translators/deveoTranslator.js
@@ -1,11 +1,5 @@
 'use strict';
 
-var _get = require('babel-runtime/helpers/get')['default'];
-
-var _inherits = require('babel-runtime/helpers/inherits')['default'];
-
-var _classCallCheck = require('babel-runtime/helpers/class-call-check')['default'];
-
 var _interopRequireDefault = require('babel-runtime/helpers/interop-require-default')['default'];
 
 Object.defineProperty(exports, '__esModule', {
@@ -20,33 +14,17 @@ var _uuidV4 = require('uuid-v4');
 
 var _uuidV42 = _interopRequireDefault(_uuidV4);
 
-var _middlewareCsError = require('../../middleware/csError');
+var _middlewareDeveoCommitMalformedError = require('../../middleware/deveoCommitMalformedError');
 
-var _middlewareCsError2 = _interopRequireDefault(_middlewareCsError);
+var _middlewareDeveoCommitMalformedError2 = _interopRequireDefault(_middlewareDeveoCommitMalformedError);
 
-var _getProperties2 = require('./getProperties');
+var _getProperties = require('./getProperties');
 
-var _getProperties3 = _interopRequireDefault(_getProperties2);
+var _getProperties2 = _interopRequireDefault(_getProperties);
 
 var _branchNameParse = require('./branchNameParse');
 
 var _branchNameParse2 = _interopRequireDefault(_branchNameParse);
-
-//TODO: do we want this kind of library to know about status codes?
-
-var DeveoCommitMalformedError = (function (_CSError) {
-  _inherits(DeveoCommitMalformedError, _CSError);
-
-  function DeveoCommitMalformedError(error, pushEvent) {
-    _classCallCheck(this, DeveoCommitMalformedError);
-
-    _get(Object.getPrototypeOf(DeveoCommitMalformedError.prototype), 'constructor', this).call(this, ['There was an unexpected error when processing your Deveo push event.']);
-    this.originalError = error;
-    this.pushEvent = pushEvent;
-  }
-
-  return DeveoCommitMalformedError;
-})(_middlewareCsError2['default']);
 
 var deveoTranslator = {
   family: 'Deveo',
@@ -64,9 +42,9 @@ var deveoTranslator = {
             var commit = {
               sha: aCommit.id,
               commit: {
-                author: pushEvent.repository.type == "subversion" ? pushEvent.pusher.display_name : aCommit.author.name,
+                author: pushEvent.repository.type == 'subversion' ? pushEvent.pusher.display_name : aCommit.author.name,
                 committer: {
-                  name: pushEvent.repository.type == "subversion" ? pushEvent.pusher.display_name : aCommit.author.name,
+                  name: pushEvent.repository.type == 'subversion' ? pushEvent.pusher.display_name : aCommit.author.name,
                   email: aCommit.author.email,
                   date: aCommit.timestamp
                 },
@@ -78,7 +56,6 @@ var deveoTranslator = {
               originalMessage: aCommit
             };
             return {
-
               eventId: (0, _uuidV42['default'])(),
               eventType: 'DeveoCommitReceived',
               data: commit,
@@ -94,15 +71,15 @@ var deveoTranslator = {
 
       if (typeof _ret === 'object') return _ret.v;
     } catch (ex) {
-      var otherEx = new DeveoCommitMalformedError(ex, pushEvent);
-      //console.log(otherEx, otherEx.originalError.stack);
-      throw otherEx;
+      throw new _middlewareDeveoCommitMalformedError2['default'](ex, pushEvent);
     }
   },
+
   canTranslate: function canTranslate(request) {
     var headers = request.headers;
     return headers.hasOwnProperty('x-deveo-event') && headers['x-deveo-event'] === 'push';
   },
+
   getProperties: function getProperties(event) {
     var commit = event.commit;
     var branch = event.branch;
@@ -123,7 +100,6 @@ var deveoTranslator = {
       props.repo = project_name + '/' + repo_name;
       props.repoHref = serverUrl[0] + company_name + '/projects/' + project_name + '/repositories/' + repo_name;
       props.branchHref = props.repoHref + '/tree/' + branch;
-
     } else {
       throw 'Could not parse DeveoCommitReceived event props correctly.';
     }
