+++ src/app/middleware/es6/deveoCommitMalformedError.js
@@ -1,0 +1,11 @@
+import CSError from './csError';
+
+class DeveoCommitMalformedError extends CSError {
+  constructor(error, pushEvent) {
+    super(['There was an unexpected error when processing your Deveo push event.']);
+    this.originalError = error;
+    this.pushEvent = pushEvent;
+  }
+}
+
+export default DeveoCommitMalformedError;
